Updated Delete Sites methods with more robust functionality
This update takes advantage of the _FormatExceptionMessage method update.
Essentially, it will help clean up a Site without having to specifically delete every object within that Site prior to.
For example, if Policies and/or Groups exist in a Site you can't delete the Site, but now, upon attempting this, the error text returns the items that are dependent on the Site to be deleted, that text is parsed and then each item is deleted individually, and then it attempts to delete the Site again.