Improved Delete Site Methods
Improved the logic in the delete Site method to verify the returned object names are _actually_ in the Site requested to be deleted.  Since it only returns the names, delete a name will delete all occurrences of that name, if the name is name unique for that object.
Also removed the logic to delete "Patch Management Software Title" as that doesn't seem to be fully flush out in the Classic API, at least in my testing so far.